+++ changelog.md
@@ -15,6 +15,7 @@
 * Edit: various changes to several plugin core classes
 * Edit: improvements to the `Features` panel filtering
 * Edit: unique library prefix for files added by `Enqueue`
+* Fix: issue with the screen options handling on networks setups
 
 ### Version 4.4 - 2023.11.06
 * New: render Changelog and History About panels from markdown